14th Commit
Barricade initialized and barricade health set. The player's experience
is now displayed at the bottom of the screen in an experience bar.
Weapons can no longer be bought for free by using the console command
(weapon names and weapon costs must now be hard-coded within the
weaponPrices array in the concommands to prevent this bug).