15th Commit
Made a level requirement of 5 for shotgun. The level requirement
argument must also be hard-coded into the weaponPrices array along with
the weapon name and price. Fixed the F4 menu so that only the player
that opens it will see the menu instead of everyone on the server seeing
the menu also.