16th Commit
If a player buys a weapon, the weapon is saved to the player's default
loadout if the server were to be restarted, if the player were to
disconnect, or if the player were to die. Added error messages for
situations when buying entities and weapons. Also added a level up
message.